+++ TSClientGen.Extensibility/Properties/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 ﻿using System.Reflection;
 
-[assembly: AssemblyVersion("9.2.1")]
-[assembly: AssemblyInformationalVersion("9.2.1")]
+[assembly: AssemblyVersion("9.2.2")]
+[assembly: AssemblyInformationalVersion("9.2.2")]
 
 [assembly: AssemblyTitle("TSClientGen.Extensibility")]
 [assembly: AssemblyDescription("Contracts assembly for authoring plugins for TSClientGen tool")]